requirements.txt
PySide6>=6.4.0
pdf2image>=1.16.3
pytesseract>=0.3.10
easyocr>=1.7.1
beautifulsoup4>=4.12.2
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
torchvision>=0.15.0 